+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.26.0 (BREAKING CHANGES)
+    * use `AsyncCommands` trait from `redis-rs` and remove own `Commands` trait
+    * add trait `RedisOps` to support specialized use cases like ping etc.
+    * reexport `redis` crate
+    * `RedisPool` implements `ConnectionLike`
+    * `RedisPool::ping` was renamed to `RedisPool::ping_nodes` because there is also a simple ping on `RedisOps`
+
 ## 0.25.0 (BREAKING CHANGE)
     * drop connections only on IO errors
     * errors are now `Sync`